+++ zproject/settings.py
@@ -159,6 +159,7 @@ DEFAULT_SETTINGS = {
     'SOCIAL_AUTH_GITHUB_TEAM_ID': None,
     'SOCIAL_AUTH_SUBDOMAIN': None,
     'SOCIAL_AUTH_AZUREAD_OAUTH2_SECRET': get_secret('azure_oauth2_secret'),
+	'SOCIAL_AUTH_FACEBOOK_SECRET': get_secret('facebook_secret'),
 
     # Other auth
     'SSO_APPEND_DOMAIN': None,
@@ -1434,3 +1435,15 @@ CROSS_REALM_BOT_EMAILS = {
 CONTRIBUTORS_DATA = os.path.join(STATIC_ROOT, 'generated/github-contributors.json')
 
 THUMBOR_KEY = get_secret('thumbor_key')
+
+########################################################################
+#
+# Facebook OAuth Settings, for developers
+# who would like specific API version/define scope
+#
+########################################################################
+
+SOCIAL_AUTH_FACEBOOK_API_VERSION = '3.2'
+SOCIAL_AUTH_FACEBOOK_APP_NAMESPACE = ''
+SOCIAL_AUTH_FACEBOOK_SCOPE = ['email']
+SOCIAL_AUTH_FACEBOOK_PROFILE_EXTRA_PARAMS = {'fields': 'id,name,email', }